content/xml/document/test/test_bug343870.xhtml

<html xmlns="http://www.w3.org/1999/xhtml">
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=343870
-->
<head>
  <title>Test for Bug 343870</title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>        
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=343870">Mozilla Bug 343870</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">
<![CDATA[

/** Test for Bug 343870 **/
function doTest() {
  var dataDoc = document.getElementById("d").contentDocument;
  is(dataDoc.getElementById("lf").firstChild.data, "\n");
  is(dataDoc.getElementById("cr").firstChild.data, "\n");
  is(dataDoc.getElementById("crlf").firstChild.data, "\n");
  is(dataDoc.getElementById("escapedcr").firstChild.data, "\r");
}

SimpleTest.waitForExplicitFinish();
addLoadEvent(doTest);
addLoadEvent(SimpleTest.finish);

]]>
</script>
</pre>
<iframe id="d" src="data:application/xml,%3C%3Fxml version%3D'1.0' encoding%3D'utf-8'%3F%3E%3Chtml xmlns%3D'http%3A//www.w3.org/1999/xhtml'%3E%3Cp id%3D'lf'%3E%0A%3C/p%3E%3Cp id%3D'cr'%3E%0D%3C/p%3E%3Cp id%3D'crlf'%3E%0D%0A%3C/p%3E%3Cp id%3D'escapedcr'%3E%26%2313;%3C/p%3E%3C/html%3E"></iframe>
</body>
</html>
